Entry point moved to main.cpp in VideoCapture and ScreenCapture projects #61
