+++ Capture/screencapture/CMakeLists.txt
@@ -2,9 +2,11 @@
 
 project(reprostim-screencapture)
 
-
 # Create the executable
-add_executable(${PROJECT_NAME} ScreenCapture.cpp)
+add_executable(${PROJECT_NAME}
+        ScreenCapture.cpp
+        main.cpp
+)
 # Copy config.yaml to out folder as well
 configure_file(${CMAKE_CURRENT_SOURCE_DIR}/config.yaml ${CMAKE_CURRENT_BINARY_DIR}/config.yaml COPYONLY)
