+++ Capture/videocapture/CMakeLists.txt
@@ -3,7 +3,10 @@
 project(reprostim-videocapture)
 
 # Create the executable
-add_executable(${PROJECT_NAME} VideoCapture.cpp)
+add_executable(${PROJECT_NAME}
+        VideoCapture.cpp
+        main.cpp
+)
 # Copy config.yaml to out folder as well
 configure_file(${CMAKE_CURRENT_SOURCE_DIR}/config.yaml ${CMAKE_CURRENT_BINARY_DIR}/config.yaml COPYONLY)
