+++ Capture/videocapture/VideoCapture.cpp
@@ -218,15 +218,3 @@ void VideoCaptureApp::stopRecording(const std::string& start_ts, const std::stri
 	}
 }
 
-////////////////////////////////////////////////////////////////////////////
-// Entry point
-
-int main(int argc, char* argv[]) {
-	int res = EX_OK;
-	do {
-		VideoCaptureApp app;
-		res = app.run(argc, argv);
-		optind = 0; // force restart argument scanning for getopt
-	} while( res==EX_CONFIG_RELOAD );
-	return res;
-}
